pyproject.toml
[project]
name = "opencode-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "black>=26.1.0",
    "pyjwt>=2.10.1",
]
